dts / zc702: incorporate TRD and rpmsg DTS nodes
The Xilinx TRD provides a dts that includes support for the logicvc and
Xylon DRM nodes, among other enhancements.  Use that as the default dts
for building zc702 images.

In order to support Linux as a remoteproc slave, we also want to provide
the rpmsg driver node, as an optional, default-enabled, included node.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>